requirements.txt
Flask==2.0.0
requests==2.25.1
python-dotenv===0.17.1
PyGithub===1.55
dropbox===11.9.0
boto3==1.17.72
pyftpdlib===1.5.6
python-gitlab===2.7.1
yagmail===0.14.256
gunicorn==20.0.4